perf: Lazy load Tooljet Database page, reduces initial chunk size by ~102.19 KB
