+++ frontend/src/AppBuilder/QueryManager/QueryEditors/TooljetDatabase/DeleteRows.jsx
@@ -1,8 +1,8 @@
-import React, { useContext } from 'react';
-import { TooljetDatabaseContext } from '@/TooljetDatabase/index';
+import React from 'react';
+import { useToolJetDbOperationsContext } from './ToolJetDbOperationsContext';
 import { v4 as uuidv4 } from 'uuid';
 import { isEmpty } from 'lodash';
-import { operators } from '@/TooljetDatabase/constants';
+import { operators } from '@/_helpers/constants';
 import { isOperatorOptions } from './util';
 import { ButtonSolid } from '@/_ui/AppButton/AppButton';
 import CodeHinter from '@/AppBuilder/CodeEditor';
@@ -11,7 +11,7 @@ import { NoCondition } from './NoConditionUI';
 
 export const DeleteRows = React.memo(({ darkMode }) => {
   const { columns, deleteOperationLimitOptionChanged, deleteRowsOptions, handleDeleteRowsOptionsChange } =
-    useContext(TooljetDatabaseContext);
+    useToolJetDbOperationsContext();
 
   function handleWhereFiltersChange(filters) {
     handleDeleteRowsOptionsChange('where_filters', filters);
