+++ frontend/src/AppBuilder/QueryManager/QueryEditors/TooljetDatabase/JoinConstraint.jsx
@@ -1,6 +1,6 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { Container, Row, Col } from 'react-bootstrap';
-import { TooljetDatabaseContext } from '@/TooljetDatabase/index';
+import { useToolJetDbOperationsContext } from './ToolJetDbOperationsContext';
 import DropDownSelect from './DropDownSelect';
 import { ButtonSolid } from '@/_ui/AppButton/AppButton';
 import AddRectangle from '@/_ui/Icon/bulkIcons/AddRectangle';
@@ -19,7 +19,7 @@ import { ToolTip } from '@/_components';
 
 const JoinConstraint = ({ darkMode, index, onRemove, onChange, data }) => {
   const { selectedTableId, tables, joinOptions, findTableDetails, tableForeignKeyInfo } =
-    useContext(TooljetDatabaseContext);
+    useToolJetDbOperationsContext();
   const joinType = data?.joinType;
   const baseTableDetails = (selectedTableId && findTableDetails(selectedTableId)) || {};
   const conditionsList = isEmpty(data?.conditions?.conditionsList) ? [{}] : data?.conditions?.conditionsList;
@@ -399,7 +399,7 @@ const JoinOn = ({
   onOperatorChange,
   onRemove,
 }) => {
-  const { tableInfo, findTableDetails } = useContext(TooljetDatabaseContext);
+  const { tableInfo, findTableDetails } = useToolJetDbOperationsContext();
   const { operator, leftField, rightField } = condition;
   const leftFieldColumn = leftField?.columnName;
   const rightFieldColumn = rightField?.columnName;
