+++ frontend/src/AppBuilder/QueryManager/QueryEditors/TooljetDatabase/ToolJetDbOperationsContext.jsx
@@ -1,0 +1,17 @@
+import React, { createContext, useContext } from 'react';
+
+const ToolJetDbOperationsContext = createContext();
+
+export const ToolJetDbOperationsProvider = ({ value, children }) => (
+  <ToolJetDbOperationsContext.Provider value={value}>{children}</ToolJetDbOperationsContext.Provider>
+);
+
+export const useToolJetDbOperationsContext = () => {
+  const context = useContext(ToolJetDbOperationsContext);
+
+  if (!context) {
+    throw new Error('useToolJetDbOperationsContext must be used within a ToolJetDbOperationsProvider');
+  }
+
+  return context;
+};
