+++ frontend/src/TooljetDatabase/Drawers/CreateTableDrawer/index.jsx
@@ -2,14 +2,15 @@ import React, { useState, useContext, useEffect } from 'react';
 import { toast } from 'react-hot-toast';
 import Drawer from '@/_ui/Drawer';
 import CreateTableForm from '../../Forms/TableForm';
-import { TooljetDatabaseContext } from '../../index';
-import { tooljetDatabaseService, authenticationService } from '@/_services';
+import { useTooljetDatabaseContext } from '../../TooljetDatabaseContext';
+import { tooljetDatabaseService } from '@/_services/tooljetDatabase.service';
+import { authenticationService } from '@/_services/authentication.service';
 import { ButtonSolid } from '@/_ui/AppButton/AppButton';
 import { BreadCrumbContext } from '@/App/App';
 import posthogHelper from '@/modules/common/helpers/posthogHelper';
 
 export default function CreateTableDrawer({ bannerVisible, setBannerVisible, tablesLimit, setTablesLimit }) {
-  const { organizationId, setSelectedTable, setTables, tables } = useContext(TooljetDatabaseContext);
+  const { organizationId, setSelectedTable, setTables, tables } = useTooljetDatabaseContext();
   const [isCreateTableDrawerOpen, setIsCreateTableDrawerOpen] = useState(false);
   const { updateSidebarNAV } = useContext(BreadCrumbContext);
   setBannerVisible(tablesLimit?.current >= tablesLimit?.total - 1 || false);
