+++ frontend/src/TooljetDatabase/Forms/FilterForm.jsx
@@ -1,14 +1,14 @@
 import './styles.scss';
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Select from '@/_ui/Select';
-import { TooljetDatabaseContext } from '../index';
-import { operators } from '../constants';
+import { useTooljetDatabaseContext } from '../TooljetDatabaseContext';
+import { operators } from '@/_helpers/constants';
 import { debounce } from 'lodash';
-import { ToolTip } from '@/_components';
+import { ToolTip } from '@/_components/ToolTip';
 import SolidIcon from '@/_ui/Icon/SolidIcons';
 
 export const FilterForm = ({ filters, setFilters, index, column = '', operator = '', value = '', generateMessage }) => {
-  const { columns, setPageCount } = useContext(TooljetDatabaseContext);
+  const { columns, setPageCount } = useTooljetDatabaseContext();
 
   const [filterInputValue, setFilterInputValue] = useState(value);
