+++ frontend/src/TooljetDatabase/Forms/ForeignKeyRelation.jsx
@@ -1,19 +1,20 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { tooljetDatabaseService } from '@/_services';
+import React, { useState } from 'react';
+import { isEmpty } from 'lodash';
+import { toast } from 'react-hot-toast';
+import { Tooltip } from 'react-tooltip';
+import cx from 'classnames';
+import { tooljetDatabaseService } from '@/_services/tooljetDatabase.service';
 import Information from '../Icons/information.svg';
 import ForeignKeyRelationIcon from '../Icons/Fk-relation.svg';
 import AddRectangle from '@/_ui/Icon/bulkIcons/AddRectangle';
 import { ButtonSolid } from '@/_ui/AppButton/AppButton';
 import Drawer from '@/_ui/Drawer';
-import { toast } from 'react-hot-toast';
 import ForeignKeyTableForm from './ForeignKeyTableForm';
 import EditIcon from '../Icons/EditColumn.svg';
-import _, { isEmpty } from 'lodash';
-import { ConfirmDialog } from '@/_components';
-import { Tooltip } from 'react-tooltip';
+import { ConfirmDialog } from '@/_components/ConfirmDialog';
 import { getColumnDataType, dataTypes } from '../constants';
-import { TooljetDatabaseContext } from '../index';
-import cx from 'classnames';
+import { useTooljetDatabaseContext } from '../TooljetDatabaseContext';
+import ChangesComponent from '../ChangesComponent';
 
 function ForeignKeyRelation({
   onMouseHoverFunction = () => {},
@@ -284,7 +285,7 @@ function ForeignKeyRelation({
 
   // const isMatchingForeignKeyColumns = checkMatchingColumnNamesInForeignKey(foreignKeyDetails, columns);
 
-  const { tables } = useContext(TooljetDatabaseContext);
+  const { tables } = useTooljetDatabaseContext();
 
   const checkTablelength = (tableLength, editMode) => {
     if (editMode) {
@@ -497,15 +498,15 @@ function ForeignKeyRelation({
         confirmButtonText={'Continue'}
         cancelButtonText={'Cancel'}
         footerStyle={footerStyle}
-        // currentPrimaryKeyIcons={currentPrimaryKeyIcons}
-        // newPrimaryKeyIcons={newPrimaryKeyIcons}
-        isEditToolJetDbTable={true}
-        foreignKeyChanges={newChangesInForeignKey}
-        existingReferencedTableName={existingReferencedTableName}
-        existingReferencedColumnName={existingReferencedColumnName}
-        currentReferencedTableName={currentReferencedTableName}
-        currentReferencedColumnName={currentReferencedColumnName}
-      />
+      >
+        <ChangesComponent
+          foreignKeyChanges={newChangesInForeignKey}
+          existingReferencedTableName={existingReferencedTableName}
+          existingReferencedColumnName={existingReferencedColumnName}
+          currentReferencedTableName={currentReferencedTableName}
+          currentReferencedColumnName={currentReferencedColumnName}
+        />
+      </ConfirmDialog>
     </>
   );
 }
