+++ frontend/src/TooljetDatabase/Table/Footer.jsx
@@ -1,8 +1,8 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import Select from '@/_ui/Select';
 import Pagination from '@/_ui/Pagination';
 import Skeleton from 'react-loading-skeleton';
-import { TooljetDatabaseContext } from '../index';
+import { useTooljetDatabaseContext } from '../TooljetDatabaseContext';
 import LeftNav from '../Icons/LeftNav.svg';
 import RightNav from '../Icons/RightNav.svg';
 import Enter from '../Icons/Enter.svg';
@@ -17,7 +17,7 @@ const Footer = ({ darkMode, dataLoading, tableDataLength, collapseSidebar }) => 
   ];
 
   const { selectedTable, totalRecords, buildPaginationQuery, setPageCount, pageCount, setPageSize, pageSize } =
-    useContext(TooljetDatabaseContext);
+    useTooljetDatabaseContext();
 
   const totalPage = Math.ceil(totalRecords / pageSize);
   const pageRange = `${(pageCount - 1) * pageSize + 1} - ${
