+++ frontend/src/_helpers/constants.js
@@ -166,3 +166,18 @@ export const PLANS = {
   TRIAL: 'trial',
   STARTER: 'starter',
 };
+
+export const operators = [
+  { value: 'eq', label: 'equals' },
+  { value: 'gt', label: 'greater than' },
+  { value: 'gte', label: 'greater than or equal' },
+  { value: 'lt', label: 'less than' },
+  { value: 'lte', label: 'less than or equal' },
+  { value: 'neq', label: 'not equal' },
+  { value: 'like', label: 'like' },
+  { value: 'ilike', label: 'ilike' },
+  { value: 'match', label: 'match' },
+  { value: 'imatch', label: 'imatch' },
+  { value: 'in', label: 'in' },
+  { value: 'is', label: 'is' },
+];
